--- a/libcxx/test/std/algorithms/alg.modifying.operations/alg.move/move.pass.cpp
+++ b/libcxx/test/std/algorithms/alg.modifying.operations/alg.move/move.pass.cpp
@@ -94,20 +94,22 @@ struct Test1OutIters {
 };
 
 TEST_CONSTEXPR_CXX20 bool test_vector_bool(std::size_t N) {
-  std::vector<bool> in(N, false);
+  std::vector<bool> v(N, false);
   for (std::size_t i = 0; i < N; i += 2)
-    in[i] = true;
+    v[i] = true;
 
   { // Test move with aligned bytes
+    std::vector<bool> in(v);
     std::vector<bool> out(N);
     std::move(in.begin(), in.end(), out.begin());
-    assert(in == out);
+    assert(out == v);
   }
   { // Test move with unaligned bytes
-    std::vector<bool> out(N + 8);
-    std::move(in.begin(), in.end(), out.begin() + 4);
-    for (std::size_t i = 0; i < N; ++i)
-      assert(out[i + 4] == in[i]);
+    std::vector<bool> in(v);
+    std::vector<bool> out(N);
+    std::move(in.begin() + 4, in.end(), out.begin());
+    for (std::size_t i = 0; i < N - 4; ++i)
+      assert(v[i + 4] == out[i]);
   }
 
   return true;