As a follow-up to #121013 (which optimized `ranges::copy`) and #121026 (which optimized `ranges::copy_backward`), this PR enhances the performance of `std::ranges::{move, move_backward}` for `vector<bool>::iterator`, addressing a subtask outlined in issue #64038. 

The optimizations bring performance improvements analogous to those achieved for the `{copy, copy_backward}` algorithms: up to **2000x** for aligned moves and **60x** for unaligned moves. Moreover, comprehensive tests covering up to 4 storage words (256 bytes) with odd and even bit sizes are provided, which validate the proposed optimizations in this patch. 

### Benchmarks 

 #### Aligned move (up to 2111x)

```
---------------------------------------------------------------------------
Benchmark                                  Before        After  Improvement
---------------------------------------------------------------------------
bm_ranges_move_vb_aligned/8               11.4 ns      5.29 ns           2x
bm_ranges_move_vb_aligned/64              92.1 ns      5.73 ns          16x
bm_ranges_move_vb_aligned/512              729 ns      5.31 ns         137x
bm_ranges_move_vb_aligned/4096            5759 ns      10.4 ns        1168x
bm_ranges_move_vb_aligned/32768          45092 ns      38.6 ns        15424
bm_ranges_move_vb_aligned/65536          91119 ns      63.5 ns        1435x
bm_ranges_move_vb_aligned/102400        141736 ns       112 ns        1266x
bm_ranges_move_vb_aligned/106496        146472 ns       116 ns        1263x
bm_ranges_move_vb_aligned/110592        153312 ns       121 ns        1267x
bm_ranges_move_vb_aligned/114688        159366 ns       125 ns        1275x
bm_ranges_move_vb_aligned/118784        164790 ns       130 ns        1267x
bm_ranges_move_vb_aligned/122880        171524 ns       136 ns        1261x
bm_ranges_move_vb_aligned/126976        175769 ns       143 ns        1229x
bm_ranges_move_vb_aligned/131072        182248 ns       175 ns        1041x
bm_ranges_move_vb_aligned/135168        188863 ns       177 ns        1067x
bm_ranges_move_vb_aligned/139264        192792 ns       165 ns        1168x
bm_ranges_move_vb_aligned/143360        194506 ns       156 ns        1247x
bm_ranges_move_vb_aligned/147456        203244 ns       160 ns        1270x
bm_ranges_move_vb_aligned/151552        207067 ns      99.4 ns        2083x
bm_ranges_move_vb_aligned/155648        213459 ns       116 ns        1840x
bm_ranges_move_vb_aligned/159744        219885 ns       162 ns        1357x
bm_ranges_move_vb_aligned/163840        228873 ns       142 ns        1612x
bm_ranges_move_vb_aligned/167936        233786 ns       160 ns        1461x
bm_ranges_move_vb_aligned/172032        236231 ns       129 ns        1831x
bm_ranges_move_vb_aligned/176128        247542 ns       119 ns        2080x
bm_ranges_move_vb_aligned/180224        249135 ns       118 ns        2111x
bm_ranges_move_vb_aligned/184320        253485 ns       123 ns        2060x
bm_ranges_move_vb_aligned/188416        258189 ns       148 ns        1745x
bm_ranges_move_vb_aligned/192512        267000 ns       207 ns        1290x
bm_ranges_move_vb_aligned/196608        269285 ns       185 ns        1456x
bm_ranges_move_vb_aligned/200704        279332 ns       238 ns        1174x
bm_ranges_move_vb_aligned/204800        286305 ns       236 ns        1213x
```

#### Aligned move_backward (up to 2233x)

```
-----------------------------------------------------------------------------------
Benchmark                                           Before       After  Improvement
-----------------------------------------------------------------------------------
bm_ranges_move_backward_vb_aligned/8               11.3 ns     5.81 ns           2x
bm_ranges_move_backward_vb_aligned/64              82.2 ns     6.15 ns          13x
bm_ranges_move_backward_vb_aligned/512              651 ns     5.59 ns         116x
bm_ranges_move_backward_vb_aligned/4096            5234 ns     10.9 ns         480x
bm_ranges_move_backward_vb_aligned/32768          42212 ns     39.0 ns        1082x
bm_ranges_move_backward_vb_aligned/65536          84501 ns     63.0 ns        1341x
bm_ranges_move_backward_vb_aligned/102400        133306 ns     62.0 ns        2150x
bm_ranges_move_backward_vb_aligned/106496        140072 ns     63.6 ns        2202x
bm_ranges_move_backward_vb_aligned/110592        147510 ns     66.4 ns        2222x
bm_ranges_move_backward_vb_aligned/114688        147312 ns     68.0 ns        2166x
bm_ranges_move_backward_vb_aligned/118784        155653 ns     70.7 ns        2202x
bm_ranges_move_backward_vb_aligned/122880        161665 ns     72.4 ns        2233x
bm_ranges_move_backward_vb_aligned/126976        164021 ns     76.6 ns        2142x
bm_ranges_move_backward_vb_aligned/131072        171836 ns     76.9 ns        2235x
bm_ranges_move_backward_vb_aligned/135168        177469 ns      128 ns        1386x
bm_ranges_move_backward_vb_aligned/139264        180642 ns     99.6 ns        1814x
bm_ranges_move_backward_vb_aligned/143360        186952 ns     90.3 ns        2070x
bm_ranges_move_backward_vb_aligned/147456        192811 ns     90.7 ns        2126x
bm_ranges_move_backward_vb_aligned/151552        201549 ns     94.3 ns        2137x
bm_ranges_move_backward_vb_aligned/155648        205928 ns      111 ns        1855x
bm_ranges_move_backward_vb_aligned/159744        210829 ns      152 ns        1387x
bm_ranges_move_backward_vb_aligned/163840        213758 ns      137 ns        1560x
bm_ranges_move_backward_vb_aligned/167936        221136 ns      153 ns        1445x
bm_ranges_move_backward_vb_aligned/172032        225363 ns      124 ns        1817x
bm_ranges_move_backward_vb_aligned/176128        230920 ns      116 ns        1991x
bm_ranges_move_backward_vb_aligned/180224        237196 ns      117 ns        2027x
bm_ranges_move_backward_vb_aligned/184320        243265 ns      123 ns        1978x
bm_ranges_move_backward_vb_aligned/188416        244984 ns      143 ns        1713x
bm_ranges_move_backward_vb_aligned/192512        252805 ns      197 ns        1283x
bm_ranges_move_backward_vb_aligned/196608        259500 ns      185 ns        1403x
bm_ranges_move_backward_vb_aligned/200704        260624 ns      231 ns        1128x
bm_ranges_move_backward_vb_aligned/204800        268815 ns      239 ns        1125x
```


 #### Unaligned move (up to 67x)

```
---------------------------------------------------------------------------
Benchmark                                  Before        After  Improvement
---------------------------------------------------------------------------
bm_ranges_move_vb_unaligned/8             6.23 ns      10.4 ns         0.6x
bm_ranges_move_vb_unaligned/64            89.3 ns      10.3 ns           9x
bm_ranges_move_vb_unaligned/512            741 ns      18.9 ns          39x
bm_ranges_move_vb_unaligned/4096          6059 ns      96.7 ns          63x
bm_ranges_move_vb_unaligned/32768        49585 ns       747 ns          66x
bm_ranges_move_vb_unaligned/262144      395276 ns      5895 ns          67x
bm_ranges_move_vb_unaligned/1048576    1567811 ns     23451 ns          67x
```

 #### Unaligned move_backward (up to 60x)

```
-----------------------------------------------------------------------------------
Benchmark                                           Before       After  Improvement
-----------------------------------------------------------------------------------
bm_ranges_move_backward_vb_unaligned/8             5.63 ns     10.2 ns         0.5x
bm_ranges_move_backward_vb_unaligned/64            88.2 ns     9.43 ns           9x
bm_ranges_move_backward_vb_unaligned/512            727 ns     18.4 ns          40x
bm_ranges_move_backward_vb_unaligned/4096          5789 ns      103 ns          56x
bm_ranges_move_backward_vb_unaligned/32768        45744 ns      835 ns          55x
bm_ranges_move_backward_vb_unaligned/262144      377009 ns     6307 ns          60x
bm_ranges_move_backward_vb_unaligned/1048576    1507192 ns    25390 ns          59x
```